// web-ui/client/gen/sales/leadStatusEnum.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Sales Cookbook API
 * Sales Cookbook API Documentation - A comprehensive sales management system
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `new` - New
* `contacted` - Contacted
* `qualified` - Qualified
* `converted` - Converted
* `disqualified` - Disqualified
 */
export type LeadStatusEnum = typeof LeadStatusEnum[keyof typeof LeadStatusEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const LeadStatusEnum = {
  new: 'new',
  contacted: 'contacted',
  qualified: 'qualified',
  converted: 'converted',
  disqualified: 'disqualified',
} as const;