// web-ui/client/gen/sales/stageEnum.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Sales Cookbook API
 * Sales Cookbook API Documentation - A comprehensive sales management system
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `qualification` - Qualification
* `proposal` - Proposal
* `negotiation` - Negotiation
* `won` - Closed Won
* `lost` - Closed Lost
 */
export type StageEnum = typeof StageEnum[keyof typeof StageEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const StageEnum = {
  qualification: 'qualification',
  proposal: 'proposal',
  negotiation: 'negotiation',
  won: 'won',
  lost: 'lost',
} as const;